requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
Django==5.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
ipython==8.35.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
nbformat==5.10.4
# numpy==1.24.3
numpy>=1.26.0,<2.2.0
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tf_keras==2.19.0
traitlets==5.14.3
typing_extensions==4.13.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
tensorflow
tensorflow_hub
imageio
opencv-python
git+https://github.com/tensorflow/docs
IPython
opencv-python
matplotlib
imageio
google-generativeai==0.4.1
mediapipe==0.10.9
protobuf>=3.20.3,<4.0.0